Drop NJTransit login and add performance
The URLs are publicly accessible and are what Transit.land uses as well
https://github.com/transitland/transitland-atlas/blob/master/feeds/nj-transit.dmfr.json

I checked if the downloads are the same with or without logins and they are.

Drop all the NJTransit specific bits including session handling.

Adds chuck_size to substancially increase download performance, overall time taken
was reduced from 30m to 12m when running wiht Septa,Patco,Delaware,NJTransit. This
was unrelated to the above but in my testing downloads above I noted that the python
requests code could use some improvements.